refactor: move interpreting statements logic
