+++ core/engine/src/module/mod.rs
@@ -190,243 +190,6 @@ impl Module {
             .to_string()
     }
 
-    /// Interpret statement from the module.
-    pub fn interpret_stmt(&mut self, stmt: Stmt, ctx: &Context) -> Result<()> {
-        match stmt {
-            Stmt::Fn(f) => {
-                debug!("Interpreting function: {}", f.name);
-                self.functions.push(StoredFunction::Function {
-                    function: f.clone(),
-                    defining_module: Arc::clone(&Arc::new(Mutex::new(self.clone()))),
-                });
-
-                if f.exported {
-                    self.exports
-                        .push((f.name.clone(), ExportType::Function(f.clone())));
-                }
-            }
-            Stmt::Use(u) => {
-                debug!("Interpreting use: {}", u.from.literal());
-
-                // Load the module as an Arc<Mutex<Module>>
-                let module = ctx
-                    .module_loader
-                    .load(&self.clone(), &u.from.literal(), ctx)
-                    .map_err(|_| {
-                        ModuleNotFoundError(u.from.literal().to_string(), u.from.span.clone())
-                    })?;
-
-                // Lock the loaded module for parsing and interpretation
-                let mut loaded_module = module.lock().expect("Failed to lock loaded module");
-
-                match loaded_module.parse() {
-                    Ok(_) => {}
-                    Err(e) => {
-                        print_diagnostic(e, Some(loaded_module.source().content()));
-                        std::process::exit(1);
-                    }
-                }
-
-                match loaded_module.interpret(ctx) {
-                    Ok(_) => {}
-                    Err(e) => {
-                        print_diagnostic(e, Some(loaded_module.source().content()));
-                        std::process::exit(1);
-                    }
-                }
-
-                // Collect the items to import
-                let imported_items: Vec<(String, &Token)> =
-                    u.items.iter().map(|i| (i.literal(), i)).collect();
-
-                for (name, item) in imported_items {
-                    match loaded_module.find_function(&name) {
-                        Some(StoredFunction::Function {
-                                 function,
-                                 defining_module,
-                             }) => {
-                            self.functions.push(StoredFunction::Function {
-                                function: function.clone(),
-                                defining_module: Arc::clone(&defining_module),
-                            });
-                        }
-                        Some(StoredFunction::Native(n)) => {
-                            self.functions.push(StoredFunction::Native(n.clone()));
-                        }
-                        None => {
-                            return Err(ImportError(name, item.span.clone()).into());
-                        }
-                    }
-                }
-            }
-            Stmt::Break(token) => {
-                debug!("Interpreting break statement");
-                return Err(PulseError::LoopBreak(token.span).into());
-            }
-            Stmt::Continue(token) => {
-                debug!("Interpreting continue statement");
-                return Err(PulseError::LoopContinue(token.span).into());
-            }
-            Stmt::While(while_stmt) => {
-                debug!("Interpreting while loop");
-                loop {
-                    self.interpret_expr(&while_stmt.condition, ctx)?;
-                    let condition_value = self.vm.pop().expect("Expected value on stack");
-
-                    let condition = match condition_value {
-                        Value::Bool(b) => b,
-                        _ => {
-                            return Err(NonBooleanCondition(
-                                "While loop condition".into(),
-                                while_stmt.condition.span(),
-                            )
-                                .into())
-                        }
-                    };
-
-                    if !condition {
-                        break;
-                    }
-
-                    self.enter_scope();
-                    let result = self.execute_block(while_stmt.block.clone(), ctx);
-                    self.exit_scope();
-
-                    match result {
-                        Ok(_) => {}
-                        Err(e) => match e.downcast::<PulseError>() {
-                            Ok(PulseError::LoopBreak(_)) => break,
-                            Ok(PulseError::LoopContinue(_)) => continue,
-                            Ok(other) => return Err(other.into()),
-                            Err(e) => return Err(e),
-                        },
-                    }
-                }
-            }
-            Stmt::Loop(loop_stmt) => {
-                debug!("Interpreting infinite loop");
-                loop {
-                    self.enter_scope();
-                    let result = self.execute_block(loop_stmt.block.clone(), ctx);
-                    self.exit_scope();
-
-                    match result {
-                        Ok(_) => {}
-                        Err(e) => match e.downcast::<PulseError>() {
-                            Ok(PulseError::LoopBreak(_)) => break,
-                            Ok(PulseError::LoopContinue(_)) => continue,
-                            Ok(other) => return Err(other.into()),
-                            Err(e) => return Err(e),
-                        },
-                    }
-                }
-            }
-            Stmt::Throw(throw) => {
-                debug!("Interpreting throw: {:?}", throw);
-
-                self.interpret_expr(&throw.value, ctx)?;
-
-                let val = self.vm.pop().unwrap();
-
-                return Err(PulseError::Throw(val.to_string(), Vec::from(self.vm.frames())).into());
-            }
-            Stmt::Try(try_stmt) => {
-                debug!("Interpreting try: {:?}", try_stmt);
-            }
-            Stmt::Let(l) => {
-                debug!("Interpreting let: {:?}", l.ident);
-                self.interpret_expr(l.initializer.as_ref(), ctx)?;
-
-                let val = self.vm.pop().unwrap();
-                let ident = l.ident.literal();
-                self.declare_variable(ident.clone(), val);
-            }
-            Stmt::Expr(expr) => {
-                debug!("Interpreting expression: {:?}", expr);
-
-                self.interpret_expr(expr.as_ref(), ctx)?;
-            }
-            Stmt::Return(r) => {
-                debug!("Interpreting return: {:?}", r);
-
-                if let Some(expr) = r.expr {
-                    self.interpret_expr(expr.as_ref(), ctx)?;
-                }
-            }
-            Stmt::Block(block) => {
-                debug!("Interpreting block statement");
-                self.execute_block(block, ctx)?;
-            }
-            Stmt::If(if_stmt) => {
-                debug!("Interpreting if statement");
-                self.interpret_if(if_stmt, ctx)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn interpret_if(&mut self, if_stmt: If, ctx: &Context) -> Result<()> {
-        self.interpret_expr(&if_stmt.condition, ctx)?;
-        let condition_value = self.vm.pop().expect("Expected value on stack");
-
-        let condition = match condition_value {
-            Value::Bool(b) => b,
-            _ => {
-                return Err(NonBooleanCondition(
-                    "If condition".into(),
-                    TextSpan::combine(vec![if_stmt.if_token.span, if_stmt.condition.span()]),
-                )
-                    .into())
-            }
-        };
-
-        if condition {
-            self.execute_block(if_stmt.then_block, ctx)?;
-        } else {
-            let mut executed = false;
-            for else_if in if_stmt.else_ifs {
-                self.interpret_expr(&else_if.condition, ctx)?;
-                let else_if_condition = self.vm.pop().expect("Expected value on stack");
-
-                let else_if_result = match else_if_condition {
-                    Value::Bool(b) => b,
-                    _ => {
-                        return Err(NonBooleanCondition(
-                            "Else if condition".into(),
-                            else_if.condition.span(),
-                        )
-                            .into())
-                    }
-                };
-
-                if else_if_result {
-                    self.execute_block(else_if.block, ctx)?;
-                    executed = true;
-                    break;
-                }
-            }
-
-            if !executed {
-                if let Some(else_block) = if_stmt.else_block {
-                    self.execute_block(else_block.block, ctx)?;
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Execute a block of statements within a new scope.
-    fn execute_block(&mut self, block: Block, ctx: &Context) -> Result<()> {
-        self.enter_scope();
-        for stmt in block.stmts {
-            self.interpret_stmt(stmt, ctx)?;
-        }
-        self.exit_scope();
-        Ok(())
-    }
-
     pub fn extract_variable_name(expr: &Expr) -> Option<String> {
         match expr {
             Expr::Variable(v) => Some(v.ident.clone()),
