+++ playground/src/main.roan
@@ -3,9 +3,7 @@ use { println } from "std::io";
 export fn main() -> int {
     let x = 10;
 
-    x /= 3;
 
-    println("{}", x);
 
     return 0;
 }
